Changed to have a common precursor asynchronous dependency that, when satisfied, kicks off two dependent asynchronous dependencies. When those are both satisfied in whatever order, a composite dependency is satisfied.
For our application, the common precursor is getting account information from the Android AccountManager. This is possibly an asynchronous task where the AccountManager shows a provided login UI, etc. When that process is complete, we can satisfy the common precursor dependency from the resulting information.

We have several dependent asynchronous dependencies: a localization manager that acquires localized strings before becoming available for use, a utility that accesses a bound service that needs to wait for the service to be completely bound before coming available, and a few others. Once all those are complete, the composite dependency is made available at the UI level (a base Fragment or Activity) so that the application can continue on with all dependencies ready to go.